backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-multipart==0.0.12
langchain==0.3.4
langchain-community==0.3.3
langchain-google-genai==2.0.1
langchain-huggingface==0.1.0
faiss-cpu==1.9.0
python-dotenv==1.0.1
pypdf==5.0.1
sentence-transformers==3.2.0
numpy==1.26.4